Changed legacy_flows to flows_legacy
