+++ bfasst/flows_legacy/ic2_synplify_conformal.py
@@ -1,27 +1,31 @@
-"""Ic2LseConformal flow"""
+"""Ic2SynplifyConformal flow"""
 
 # pylint: disable=duplicate-code
 
-from bfasst.legacy_flows.flow import Flow
-from bfasst.legacy_flows.sub_flows.conformal import Conformal
-from bfasst.legacy_flows.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
+from bfasst.flows_legacy.flow import Flow
+from bfasst.flows_legacy.sub_flows.conformal import Conformal
+from bfasst.flows_legacy.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
 from bfasst.job import Job
-from bfasst.synth.ic2_lse import Ic2LseSynthesisTool
+from bfasst.synth.ic2_synplify import Ic2SynplifySynthesisTool
 from bfasst.types import ToolType
 
 
-class Ic2LseConformal(Flow):
-    """Ic2LseConformal flow"""
+class Ic2SynplifyConformal(Flow):
+    """Ic2SynplifyConformal flow"""
 
     def create(self):
-        """Synthesize and implement using Icecube2 lse, and compare with Conformal"""
+        """
+        Icecube2 Synplify synthesis and implementation, reverse with
+        icestorm, compare with Conformal.
+        """
         # Reset job list in case this flow is called multiple times
         self.job_list = []
 
-        lse_synth_tool = Ic2LseSynthesisTool(
+        # Run Icecube2 Synplify synthesis
+        synplify_synth_tool = Ic2SynplifySynthesisTool(
             self.design.build_dir, self.design, self.flow_args[ToolType.SYNTH]
         )
-        curr_job = Job(lse_synth_tool.create_netlist, self.design.rel_path)
+        curr_job = Job(synplify_synth_tool.create_netlist, self.design.rel_path)
         self.job_list.append(curr_job)
 
         # Run icecube implementation and icestorm bitstream reversal
@@ -30,7 +34,7 @@ class Ic2LseConformal(Flow):
         impl_and_rev_sub_flow.modify_first_job_dependencies({self.job_list[-1].uuid})
         self.job_list.extend(impl_and_rev_sub_flow.job_list)
 
-        # Set paths for conformal
+        # Set paths for Conformal
         self.design.netlist_path = self.design.build_dir / (self.design.top + "_impl.v")
         if self.design.cur_error_flow_name is None:
             self.design.reversed_netlist_path = self.design.build_dir / (
@@ -43,6 +47,7 @@ class Ic2LseConformal(Flow):
 
         # Run conformal
         curr_job = Job(self.adjust_design_object, self.design.rel_path, {self.job_list[-1].uuid})
+        self.job_list.append(curr_job)
 
         conformal_sub_flow = Conformal(self.design, self.flow_args)
         conformal_sub_flow.create()
