+++ bfasst/flows_legacy/synplify_ic2_icestorm_onespin.py
@@ -3,8 +3,8 @@
 # pylint: disable=duplicate-code
 
 from bfasst.compare.onespin import OneSpinCompareTool
-from bfasst.legacy_flows.flow import Flow
-from bfasst.legacy_flows.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
+from bfasst.flows_legacy.flow import Flow
+from bfasst.flows_legacy.sub_flows.ic2_impl_and_ice_rev import Ic2ImplAndIceRev
 from bfasst.job import Job
 from bfasst.synth.ic2_synplify import Ic2SynplifySynthesisTool
 from bfasst.types import ToolType
